openai + elevenlabs + pipecat + streaming/requirements.txt
# Voice Chat Streaming - Modular Architecture
# Core dependencies for the voice chat streaming application

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# OpenAI integration
openai>=1.3.0

# HTTP client for API calls
httpx>=0.25.0

# Environment and configuration
python-dotenv>=1.0.0
pipecat
# Logging
loguru>=0.7.2

# Audio processing (optional but recommended)
# pydub>=0.25.1

# Pipecat for streaming (optional - install manually if needed)
# pipecat-ai>=0.0.1

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.9.0
# isort>=5.12.0

# Additional audio dependencies (install if needed)
# torch>=2.0.0
# torchaudio>=2.0.0
# transformers>=4.30.0
# openai-whisper>=20231117

# System audio (install if needed for advanced features)
# pyaudio>=0.2.11
# sounddevice>=0.4.6

# Note: Some dependencies may require system-level installations:
# - On macOS: brew install portaudio
# - On Ubuntu: sudo apt-get install portaudio19-dev
# - On Windows: Usually works out of the box